requirements.txt
Django==1.11.5
django-sha2==0.4
django-tastypie==0.14.0
python-decouple==3.1
gunicorn==19.6.0
whitenoise==3.3.0
django-fernet-fields==0.5
psycopg2==2.7.1
dj-database-url==0.4.2
dj-static==0.0.6